backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
requests==2.31.0
pandas==2.2.0
geopandas==0.14.3
shapely==2.0.2
python-multipart==0.0.9
cachetools==5.3.2
openpyxl==3.1.2
python-dotenv==1.2.2
sqlalchemy==2.0.27
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
